+++ minhas_financias/src/main/java/com/example/minhas_financias/model/repository/LancamentoRepository.java
@@ -13,12 +13,12 @@ import com.example.minhas_financias.model.entity.Usuario;
 import com.example.minhas_financias.model.enuns.TipoLancamento;
 
 @Repository
-public interface LancamentoRepository extends JpaRepository<Lancamento,Long> {
+public interface LancamentoRepository extends JpaRepository<Lancamento,Integer> {
 
     List<Lancamento> findByUsuario(Usuario usuario);
     
     @Query(value = "SELECT SUM(l.valor) FROM Lancamento l JOIN l.usuario u " +
             "WHERE u.id = :idUsuario AND l.tipo = :tipo")
-    BigDecimal obterSaldoLancamentoUsuario(@Param("idUsuario") Long idUsuario,@Param("tipo") TipoLancamento tipo);
+    BigDecimal obterSaldoLancamentoUsuario(@Param("idUsuario") Integer idUsuario,@Param("tipo") TipoLancamento tipo);
 
 }
